# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Concessionária API",
        "description": "API REST para gerenciamento de veículos e leads de uma concessionária.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/vehicles": {
            "get": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Listar veículos",
                "description": "Retorna uma lista paginada de veículos com suporte a filtros.",
                "operationId": "d17171521b86a0b324bfa4d4f6ca1f8a",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Tipo do veículo",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "car",
                                "motorcycle"
                            ]
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Preço máximo",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de veículos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Vehicle"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 45
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Criar veículo",
                "description": "Cria um novo veículo. Requer autenticação.",
                "operationId": "e19336f3d4ff20d9858a84d5113213bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "model",
                                    "year",
                                    "price",
                                    "color",
                                    "mileage"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "car",
                                        "enum": [
                                            "car",
                                            "motorcycle"
                                        ]
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Honda"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "Civic"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2024
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 150000
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Prata"
                                    },
                                    "mileage": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Veículo criado com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicles/{id}": {
            "get": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Detalhes do veículo",
                "description": "Retorna os detalhes de um veículo específico.",
                "operationId": "10887c4dacfad2ebc95acc8bcf1b7188",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do veículo"
                    },
                    "404": {
                        "description": "Veículo não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Atualizar veículo",
                "description": "Atualiza os dados de um veículo. Requer autenticação.",
                "operationId": "5a154eaaf029bc3a95ff0cac19946fc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "car",
                                            "motorcycle"
                                        ]
                                    },
                                    "brand": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "mileage": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Veículo atualizado com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Veículo não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Remover veículo",
                "description": "Remove um veículo do sistema. Requer autenticação.",
                "operationId": "6c00c88c53d96dd1f08e69ba6d6484a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Veículo removido com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Veículo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Vehicle": {
                "title": "Veículo",
                "description": "Modelo de veículo da concessionária",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "car",
                        "enum": [
                            "car",
                            "motorcycle"
                        ]
                    },
                    "brand": {
                        "type": "string",
                        "example": "Honda"
                    },
                    "model": {
                        "type": "string",
                        "example": "Civic"
                    },
                    "year": {
                        "type": "integer",
                        "example": 2024
                    },
                    "price": {
                        "type": "number",
                        "example": 150000
                    },
                    "color": {
                        "type": "string",
                        "example": "Prata"
                    },
                    "mileage": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}